Implement recommendations from the PR review
