+++ app/src/androidTest/java/com/github/h3lp3rs/h3lp/signin/SignedInUserTest.kt
@@ -1,11 +1,11 @@
-package com.github.h3lp3rs.h3lp.signIn
+package com.github.h3lp3rs.h3lp.signin
 
 import android.content.Intent
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.matcher.IntentMatchers
+import androidx.test.espresso.intent.Intents.*
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,11 +19,11 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito
-import org.mockito.Mockito.`when`
 import org.mockito.kotlin.anyOrNull
+import org.mockito.Mockito.`when` as When
 
 @RunWith(AndroidJUnit4::class)
-class SignInActivitySignedInUserTest {
+class SignedInUserTest {
 
     @get:Rule
     val testRule = ActivityScenarioRule(
@@ -32,18 +32,18 @@ class SignInActivitySignedInUserTest {
 
     @Before
     fun setUp() {
-        Intents.init()
+        init()
 
         val signInMock = Mockito.mock(SignInInterface::class.java)
-        `when`(signInMock.isSignedIn()).thenReturn(true)
+        When(signInMock.isSignedIn()).thenReturn(true)
 
         testRule.scenario.onActivity { activity ->
             val intent = Intent(ApplicationProvider.getApplicationContext(), activity.javaClass)
             val taskMock = Mockito.mock(Task::class.java)
-            `when`(taskMock.isSuccessful).thenReturn(true)
-            `when`(taskMock.isComplete).thenReturn(true)
-            `when`(signInMock.signIn(activity)).thenReturn(intent)
-            `when`(signInMock.authenticate(anyOrNull(), anyOrNull())).thenReturn(taskMock)
+            When(taskMock.isSuccessful).thenReturn(true)
+            When(taskMock.isComplete).thenReturn(true)
+            When(signInMock.signIn(activity)).thenReturn(intent)
+            When(signInMock.authenticate(anyOrNull(), anyOrNull())).thenReturn(taskMock)
         }
 
         SignIn.set(signInMock as SignInInterface<AuthResult>)
@@ -52,11 +52,11 @@ class SignInActivitySignedInUserTest {
     @Test
     fun signedInUserMovesToMainPageDirectly() {
         onView(withId(R.id.signInButton)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(MainPageActivity::class.java.name))
+        intended(hasComponent(MainPageActivity::class.java.name))
     }
 
     @After
     fun cleanUp() {
-        Intents.release()
+        release()
     }
 }
